requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.1
django-ckeditor==6.2.0
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.1.0
Pillow==9.0.0
psycopg2==2.9.3
python-decouple==3.5
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
